fix(pytorch) Add correct include directories for hipRTC
The hipCIM team found an error during testing related to failed
JIT compilation. The error was related to pytorch. AMD pytorch
includes libhiprtc.so in its package, and the .so is loaded after
importing pytorch.

When cupy is imported after pytorch, cupy detects that libhiprtc.so
has been loaded, and switches the JIT backend from hipcc to hiprtc.
If ROCM_HOME is not set, then the hiprtc backend is broken as the
include directories paths (like /opt/rocm/include) are not passed
as an option to hiprtc.

When ROCM_HOME is not set, this commit uses whatever hipcc is in the
PATH, and queries it's default system paths. This way, wherever cupy
is installed, and for whichever ROCm is installed, regardless of ROCM_HOME
set or not, the correct directories should be included.

Moved the driver build version check after the ROCM_HOME check. This is
because `options += ('-I' + get_rocm_path() + '/llvm/lib/clang/13.0.0/include/'`
is included in
`options += tuple(f"-I{include_dir}" for include_dir in _get_hipcc_include_dirs())`,
so it is not necessary to add twice.