feat(tests) Enable test_install_library
Test install library is currently broken. The call to `rec[assets][some_key]`
expects `some_key` to be in the form `Linux:x86_64`. Additionally, the
install_lib function expects four arguments, including the system arch.

Lastly, there is a deprecation warning for python 3.14. Since we are not
yet releasing for python 3.14, we mark these tests as expected warnings.